+++ libpyosdev/math.py
@@ -1,0 +1,2 @@
+def realmode_physaddr(offset: int, segment: int):
+    return offset * 16 + segment